chore: move confirmation utils from shared/modules to shared/lib
