+++ shared/lib/confirmation.utils.ts
