+++ ui/pages/confirmations/hooks/alerts/transactions/usePendingTransactionAlerts.ts
@@ -8,7 +8,7 @@ import { Alert } from '../../../../../ducks/confirm-alerts/confirm-alerts';
 import { Severity } from '../../../../../helpers/constants/design-system';
 import { RowAlertKey } from '../../../../../components/app/confirm/info/row/constants';
 import { useConfirmContext } from '../../../context/confirm';
-import { isCorrectDeveloperTransactionType } from '../../../../../../shared/modules/confirmation.utils';
+import { isCorrectDeveloperTransactionType } from '../../../../../../shared/lib/confirmation.utils';
 
 export function usePendingTransactionAlerts(): Alert[] {
   const t = useI18nContext();
