+++ ui/pages/confirmations/hooks/alerts/transactions/useSigningOrSubmittingAlerts.ts
@@ -7,7 +7,7 @@ import { Severity } from '../../../../../helpers/constants/design-system';
 import { useI18nContext } from '../../../../../hooks/useI18nContext';
 import { Alert } from '../../../../../ducks/confirm-alerts/confirm-alerts';
 import { useConfirmContext } from '../../../context/confirm';
-import { isCorrectDeveloperTransactionType } from '../../../../../../shared/modules/confirmation.utils';
+import { isCorrectDeveloperTransactionType } from '../../../../../../shared/lib/confirmation.utils';
 
 export function useSigningOrSubmittingAlerts(): Alert[] {
   const t = useI18nContext();
