+++ ui/pages/confirmations/hooks/alerts/useBlockaidAlerts.ts
@@ -16,7 +16,7 @@ import { Alert } from '../../../../ducks/confirm-alerts/confirm-alerts';
 import ZENDESK_URLS from '../../../../helpers/constants/zendesk-url';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
 import { SIGNATURE_TRANSACTION_TYPES } from '../../utils';
-import { isCorrectDeveloperTransactionType } from '../../../../../shared/modules/confirmation.utils';
+import { isCorrectDeveloperTransactionType } from '../../../../../shared/lib/confirmation.utils';
 import {
   SecurityAlertResponse,
   SignatureRequestType,
